refactor(#164): start annotations/ subpackage — sections.py (P5a)
Begin splitting annotate.py (2587 lines) into the annotations/ subpackage. P5a
moves the section A-A and detail views — _add_section_view, _add_detail_view, and
their helpers _section_hatch_edges (ISO 128-50 hatching) and _fuzzy_cut — into
annotations/sections.py.

The submodule sits below annotate.py: it imports only _core + build123d/OCP +
build123d_drafting, never annotate/make_drawing, so the orchestrator
(_auto_annotate) calls down into it with no cycle. annotate.py re-imports the two
pass functions; the _section_hatch_edges test now imports from the new home.

Pure move: golden gate passes against the committed snapshots with no
regeneration; section/detail/hatch + section_aa tests pass; ruff check + format +
mypy clean. Adversarially checked: no upward import statement, no stale defs,
section-view path still sections + hatches.

Part of #164 (annotations/ split; P5b-P5e follow).
